Fix Artem's remark
